+++ projects/ng2-pdf-viewer/src/public-api.ts
@@ -1,2 +1,3 @@
 export * from './lib/pdf-viewer/pdf-viewer.component';
+export * from './lib/pdf-viewer/pdf-viewer.config';
 export * from './lib/pdf-viewer/typings';